requirements.txt
flask
gunicorn
smartsheet-python-sdk
python-dotenv
